# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Challenge Corrézien Disc-Golf API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface LoginRequest {
  username: string;
  password: string;
}

export interface LoginResponse {
  success: boolean;
  message: string;
}

export interface AdminSession {
  isAuthenticated: boolean;
  username?: string;
}

export interface Year {
  id: number;
  year: number;
  createdAt: string;
}

export interface CreateYearBody {
  year: number;
}

export interface Player {
  id: number;
  name: string;
  createdAt: string;
}

export interface CreatePlayerBody {
  name: string;
}

export interface Stage {
  id: number;
  yearId: number;
  name: string;
  location?: string;
  date?: string;
  par: number;
  order: number;
  createdAt: string;
}

export interface CreateStageBody {
  yearId: number;
  name: string;
  location?: string;
  date?: string;
  par: number;
  order: number;
}

export interface Score {
  id: number;
  playerId: number;
  stageId: number;
  score: number;
  points: number;
  playerName?: string;
  stageName?: string;
  createdAt: string;
}

export interface CreateScoreBody {
  playerId: number;
  stageId: number;
  score: number;
}

export interface ScoringRule {
  id: number;
  position: number;
  points: number;
}

export interface CreateScoringRuleBody {
  position: number;
  points: number;
}

export interface StageRankingEntry {
  position: number;
  playerId: number;
  playerName: string;
  score: number;
  diffToPar: number;
  points: number;
}

export interface StageScore {
  stageId: number;
  stageName: string;
  score: number;
  par: number;
  diffToPar: number;
  points: number;
}

export interface PlayerRankingRow {
  position: number;
  playerId: number;
  playerName: string;
  stageScores: StageScore[];
  totalScore: number;
  totalPar: number;
  totalDiffToPar: number;
  totalPoints: number;
}

export interface CumulatedRankingRow {
  position: number;
  playerId: number;
  playerName: string;
  totalScore: number;
  totalPar: number;
  totalDiffToPar: number;
  totalPoints: number;
}

export type RankingsStageRankings = { [key: string]: StageRankingEntry[] };

export interface Rankings {
  yearId: number;
  year: number;
  stages: Stage[];
  stageRankings: RankingsStageRankings;
  playerRankings: PlayerRankingRow[];
  cumulatedRankings: CumulatedRankingRow[];
}

export type ListStagesParams = {
  yearId?: number;
};

export type ListScoresParams = {
  stageId?: number;
  playerId?: number;
};