# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Challenge Corrézien Disc-Golf API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current admin session
      responses:
        "200":
          description: Current session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /players:
    get:
      operationId: listPlayers
      tags: [players]
      summary: List all players
      responses:
        "200":
          description: List of players
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Player"
    post:
      operationId: createPlayer
      tags: [players]
      summary: Create a player
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlayerBody"
      responses:
        "201":
          description: Player created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
  /players/{playerId}:
    put:
      operationId: updatePlayer
      tags: [players]
      summary: Update a player
      parameters:
        - name: playerId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlayerBody"
      responses:
        "200":
          description: Player updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
    delete:
      operationId: deletePlayer
      tags: [players]
      summary: Delete a player
      parameters:
        - name: playerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Player deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /scores:
    get:
      operationId: listScores
      tags: [scores]
      summary: List scores
      parameters:
        - name: stageId
          in: query
          required: false
          schema:
            type: integer
        - name: playerId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of scores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Score"
    post:
      operationId: createOrUpdateScore
      tags: [scores]
      summary: Create or update a score
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScoreBody"
      responses:
        "200":
          description: Score saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Score"
  /scores/{scoreId}:
    delete:
      operationId: deleteScore
      tags: [scores]
      summary: Delete a score
      parameters:
        - name: scoreId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Score deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /rankings/{yearId}:
    get:
      operationId: getRankings
      tags: [rankings]
      summary: Get rankings for a year (stage breakdown + cumulated)
      parameters:
        - name: yearId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rankings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rankings"
  /rankings/{yearId}/export:
    get:
      operationId: exportRankingsCsv
      tags: [rankings]
      summary: Export rankings as CSV
      parameters:
        - name: yearId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: CSV export
          content:
            text/csv:
              schema:
                type: string
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
    AdminSession:
      type: object
      properties:
        isAuthenticated:
          type: boolean
        username:
          type: string
      required:
        - isAuthenticated
...
    Player:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - createdAt
    CreatePlayerBody:
      type: object
      properties:
        name:
          type: string
      required:
        - name
    Stage:
      type: object
      properties:
        id:
          type: integer
        yearId:
          type: integer
        name:
          type: string
        location:
          type: string
        date:
          type: string
        par:
          type: integer
        order:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - yearId
        - name
        - par
        - order
        - createdAt
...
    Score:
      type: object
      properties:
        id:
          type: integer
        playerId:
          type: integer
        stageId:
          type: integer
        score:
          type: integer
        points:
          type: integer
        playerName:
          type: string
        stageName:
          type: string
        createdAt:
          type: string
      required:
        - id
        - playerId
        - stageId
        - score
        - points
        - createdAt
    CreateScoreBody:
      type: object
      properties:
        playerId:
          type: integer
        stageId:
          type: integer
        score:
          type: integer
      required:
        - playerId
        - stageId
        - score
...
    StageRankingEntry:
      type: object
      properties:
        position:
          type: integer
        playerId:
          type: integer
        playerName:
          type: string
        score:
          type: integer
        diffToPar:
          type: integer
        points:
          type: integer
      required:
        - position
        - playerId
        - playerName
        - score
        - diffToPar
        - points
    PlayerRankingRow:
      type: object
      properties:
        position:
          type: integer
        playerId:
          type: integer
        playerName:
          type: string
        stageScores:
          type: array
          items:
            $ref: "#/components/schemas/StageScore"
        totalScore:
          type: integer
        totalPar:
          type: integer
        totalDiffToPar:
          type: integer
        totalPoints:
          type: integer
      required:
        - position
        - playerId
        - playerName
        - stageScores
        - totalScore
        - totalPar
        - totalDiffToPar
        - totalPoints
    StageScore:
      type: object
      properties:
        stageId:
          type: integer
        stageName:
          type: string
        score:
          type: integer
        par:
          type: integer
        diffToPar:
          type: integer
        points:
          type: integer
      required:
        - stageId
        - stageName
        - score
        - par
        - diffToPar
        - points
    CumulatedRankingRow:
      type: object
      properties:
        position:
          type: integer
        playerId:
          type: integer
        playerName:
          type: string
        totalScore:
          type: integer
        totalPar:
          type: integer
        totalDiffToPar:
          type: integer
        totalPoints:
          type: integer
      required:
        - position
        - playerId
        - playerName
        - totalScore
        - totalPar
        - totalDiffToPar
        - totalPoints
    Rankings:
      type: object
      properties:
        yearId:
          type: integer
        year:
          type: integer
        stages:
          type: array
          items:
            $ref: "#/components/schemas/Stage"
        stageRankings:
          type: object
          additionalProperties:
            type: array
            items:
              $ref: "#/components/schemas/StageRankingEntry"
        playerRankings:
          type: array
          items:
            $ref: "#/components/schemas/PlayerRankingRow"
        cumulatedRankings:
          type: array
          items:
            $ref: "#/components/schemas/CumulatedRankingRow"
      required:
        - yearId
        - year
        - stages
        - stageRankings
        - playerRankings
        - cumulatedRankings